+++ src/algorithm/FlashSort.cpp
@@ -1,33 +1,35 @@
 #include "algorithm.h"
 #include <cmath>
 
-void flashSort(std::vector<int> &arr, int &countComparison) 
+void flashSort(std::vector<int> &arr, unsigned long long &countComparison)
 {
     int n = arr.size();
     int min = arr[0], max = arr[0];
 
     for (int i = 0; ++countComparison && i < n; i++)
     {
-        if (++countComparison && min > arr[i]) min = arr[i];
-        if (++countComparison && max < arr[i]) max = arr[i]; 
+        if (++countComparison && min > arr[i])
+            min = arr[i];
+        if (++countComparison && max < arr[i])
+            max = arr[i];
     }
 
-    if (++countComparison && min == max) return;
+    if (++countComparison && min == max)
+        return;
 
     int m = 0.45 * n;
     std::vector<int> l(m, 0);
 
-    for (int i = 0; ++countComparison && i < n; i++) 
+    for (int i = 0; ++countComparison && i < n; i++)
     {
         int k = floor((m - 1) * (arr[i] - min) * 1.0 / (max - min));
         l[k]++;
     }
 
-    for (int i = 0; ++countComparison && i < m; i++) 
+    for (int i = 0; ++countComparison && i < m; i++)
     {
         l[i] += l[i - 1];
     }
-        
 
     int move = 0;
     int flash = 0;
@@ -35,18 +37,19 @@ void flashSort(std::vector<int> &arr, int &countComparison)
     int k = 0;
     int i = 0;
 
-    while (++countComparison && move < n - 1) 
+    while (++countComparison && move < n - 1)
     {
-        while (++countComparison && i > l[k] - 1) 
+        while (++countComparison && i > l[k] - 1)
         {
             i++;
             k = floor((m - 1) * (arr[i] - min) / (max - min));
         }
 
         flash = arr[i];
-        if (++countComparison && k < 0) break;
+        if (++countComparison && k < 0)
+            break;
 
-        while (++countComparison && i != l[k]) 
+        while (++countComparison && i != l[k])
         {
             k = floor((m - 1) * (flash - min) / (max - min));
             --l[k];
@@ -56,6 +59,6 @@ void flashSort(std::vector<int> &arr, int &countComparison)
             move++;
         }
     }
-    
+
     insertionSort(arr, countComparison);
 }
