+++ src/algorithm/MergeSort.cpp
@@ -1,52 +1,52 @@
 #include "algorithm.h"
 
-void merge(std::vector<int> &arr, int left, int mid, int right, int& countComparison) 
+void merge(std::vector<int> &arr, int left, int mid, int right, unsigned long long &countComparison)
 {
     int n = mid - left + 1;
     int m = right - mid;
 
     std::vector<int> b(n), c(m);
 
-    for (int i = 0; ++countComparison && i < n; i++) 
+    for (int i = 0; ++countComparison && i < n; i++)
     {
-        b[i] = arr[left + i]; 
+        b[i] = arr[left + i];
     }
 
-    for (int i = 0; ++countComparison && i < m; i++) 
+    for (int i = 0; ++countComparison && i < m; i++)
     {
         c[i] = arr[mid + i + 1];
     }
 
     int i = 0, j = 0, k = left;
 
-    while (++countComparison && i < n && ++countComparison && j < m) 
+    while (++countComparison && i < n && ++countComparison && j < m)
     {
-        if (++countComparison && b[i] < c[j]) 
+        if (++countComparison && b[i] < c[j])
         {
             arr[k++] = b[i++];
-        } 
-        else 
+        }
+        else
         {
             arr[k++] = c[j++];
         }
     }
 
-    while (++countComparison && i < n) 
+    while (++countComparison && i < n)
     {
         arr[k++] = b[i++];
     }
 
-    while (++countComparison && j < m) 
+    while (++countComparison && j < m)
     {
-        countComparison++; 
+        countComparison++;
         arr[k++] = c[j++];
     }
 }
 
-void mergeSortRecur(std::vector<int>& arr, int left, int right, int& countComparison) 
+void mergeSortRecur(std::vector<int> &arr, int left, int right, unsigned long long &countComparison)
 {
-    if (++countComparison && left < right) 
-    { 
+    if (++countComparison && left < right)
+    {
         int mid = (left + right) / 2;
 
         mergeSortRecur(arr, left, mid, countComparison);
@@ -56,7 +56,7 @@ void mergeSortRecur(std::vector<int>& arr, int left, int right, int& countCompar
     }
 }
 
-void mergeSort(std::vector<int>& arr, int& countComparison) 
+void mergeSort(std::vector<int> &arr, unsigned long long &countComparison)
 {
     int left = 0;
     int right = arr.size() - 1;
